requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==5.0.3
django-extensions==3.2.3
djangorestframework==3.15.1
executing==2.0.1
idna==3.6
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
mixins==0.1.4
parso==0.8.3
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
